Add files via upload
Task 1 – Start of the day.
Write a program to set up the screen display for the start of the day. Initialize suitable
data structure(s) to total passengers for each train journey and total the money taken
for each train journey. Each train journey must be totaled separately. There are four
journeys up and four journeys down every day   Task 2 – Purchasing tickets.
Tickets can be purchased for a single passenger or a group. When making a purchase,
check that the number of tickets for the required train journeys up and down the
mountain is available. If the tickets are available, calculate the total price including
any group discount. Update the screen display and the data for the totals.
Task 3 – End of the day.
Display the number of passengers that traveled on each train journey and the total
money taken for each train journey. Calculate and display the total number of
passengers and the total amount of money taken for the day. Find and display the
train journey with the most passengers that day